comentarios y permisos de edicion

+++ blog/views.py
@@ -5,13 +5,20 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
 from .models import Post, Category, Comment
 from .form import CatForm, PostForm, CommentForm
+# from accounts.utils import has_admin
+
+from django.contrib.auth.decorators import login_required, user_passes_test
 # Create your views here.
 
 def post_list(request):
-    post = Post.objects.all()
+    post = Post.objects.filter(active=True)
     return render(request, 'blog/blog_list.html', {'post': post})
 
+@login_required
+# @user_passes_test(has_admin)
 def post_create(request):
+    if not request.user.is_admin:
+        return redirect('index')
     form = PostForm()
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES)
@@ -25,7 +32,10 @@ def post_create(request):
     form = PostForm()
     return render(request, 'blog/blog_create.html', {'form': form})
 
+@login_required
 def post_update(request, pk):
+    if not request.user.is_admin:
+        return redirect('index')
     post = get_object_or_404(Post, id=pk)
     data = post
     form = PostForm(initial={'title':data.title, 'content':data.content, 'active':data.active, 'category':data.category})
@@ -36,13 +46,19 @@ def post_update(request, pk):
             return redirect('post_list')
     return render(request, 'blog/blog_update.html', {'form': form, 'post': post})
 
+@login_required
 def post_delete(request, pk):
+    if not request.user.is_admin:
+        return redirect('index')
     post = get_object_or_404(Post, id=pk)
     post.delete()
     return redirect('post_list')
 
 def post_view(request, slug):
     post = get_object_or_404(Post, slug=slug)
+    if not request.user.is_admin:
+        post.num_views = post.num_views + 1
+    post.save()
     # post = Post.objects.all(id=pk)
     return render(request, 'blog/blog_view.html', {'post': post})
 
@@ -58,7 +74,14 @@ def category_view(request):
     form = CatForm()
     return render(request, 'blog/allcategory.html', {'cat': cat, 'form': form})
 
+def post_inactive(request):
+    post = Post.objects.filter(active=False)
+    return render(request, 'blog/blog_inact.html', {'post':post})
+
+@login_required
 def category_update(request, pk):
+    if not request.user.is_admin:
+        return redirect('index')
     cat = Category.objects.all()
     cat_edit = get_object_or_404(Category, id=pk)
     data = cat_edit
@@ -70,16 +93,20 @@ def category_update(request, pk):
             return redirect('category_view')
     return render(request, 'blog/allcategory_edit.html', {'form': form, 'cat': cat, 'cat_edit': cat_edit, 'data': data})
 
+@login_required
 def cat_delete(request, pk):
+    if not request.user.is_admin:
+        return redirect('index')
     cat = get_object_or_404(Category, id=pk)
     cat.delete()
     return redirect('category_view')
 
 def categoris(request, pk):
-    post = Post.objects.filter(category__id=pk)
+    post = Post.objects.filter(category__id=pk, active=True)
     # post = Post.objects.all(category.id=pk)
     return render(request, 'blog/blog_cate.html', {'post':post})
 
+@login_required
 def comment_create(request, slug):
     context = {}
     post = get_object_or_404(Post, slug=slug)
@@ -94,16 +121,24 @@ def comment_create(request, slug):
             comment.user = request.user
             comment.save()
             return redirect('post_view', slug=post.slug)
-    return render(request, 'blog/comment_create.html', context)
+    return render(request, 'blog/blog_view.html', context)
 
+@login_required
 def comment_delete(request, pk):
-    comment = get_object_or_404(Comment, id=pk)
+    if request.user.is_admin:
+        comment = get_object_or_404(Comment, id=pk)
+    else:
+        comment = get_object_or_404(Comment, id=pk, user=request.user)
     comment.delete()
     return redirect('post_view', slug=comment.post.slug)
 
+@login_required
 def comment_update(request, pk):
     context = {}
-    comment = get_object_or_404(Comment, id=pk)
+    if request.user.is_admin:
+        comment = get_object_or_404(Comment, id=pk)
+    else:
+        comment = get_object_or_404(Comment, id=pk, user=request.user)
     data = comment
     context['commen'] = comment
     form = CommentForm(initial={'content':data.content})
@@ -117,12 +152,3 @@ def comment_update(request, pk):
             return redirect('post_view', slug=post.slug)
     return render(request, 'blog/comment_update.html', context)
     
-    # comment = get_object_or_404(Comment, id=pk)
-    # data = comment
-    # form = CommentForm(initial={'content':data.content})
-    # if request.method == 'POST':
-    #     form = CommentForm(request.POST, instance=comment)
-    #     if form.is_valid():
-    #         form.save()
-    #         return redirect('post_view', slug=comment.post.slug)
-    # return render(request, 'blog/comment_update.html', {'form': form})
